feat: Add collapsible sections for Hemograma and Glicemia exams in ExamsDetected screen
