+++ src/app/screens/ExamsHistory/index.tsx
@@ -27,8 +27,6 @@ type Row = {
 
 // TODO: fazer o map que acabe com os códigos repetidos
 export default function ExamsHistory({ navigation }: any) {
-    const [showExamsDetected, setShowExamsDetected] = useState(false);
-
     type ItemName = "glucose" | "hemogram" | "cholesterol" | "ast" | "alt";
 
     const [collapsedItems, setCollapsedItems] = useState({
@@ -227,212 +225,6 @@ export default function ExamsHistory({ navigation }: any) {
                     </ButtonContainer>
                 </Button>
             </View>
-
-            {/* {showExamsDetected ? (
-                <ModalContainer>
-                    <Header
-                        text="Exames Detectados"
-                        isModal
-                        onModalClose={() => {
-                            setShowExamsDetected(false);
-                            setCollapsedItems({
-                                glucose: true,
-                                hemogram: true,
-                                cholesterol: true,
-                                ast: true,
-                                alt: true
-                            });
-                        }}
-                    />
-                    <View>
-                        <ListInteractableItem
-                            text="Hemograma"
-                            downChevron
-                            changeIcon={collapsedItems.hemogram}
-                            isButton
-                            modalFunction={() => toggleCollapsible("hemogram")}
-                        />
-
-                        <Collapsible
-                            collapsed={collapsedItems.hemogram}
-                            align="center"
-                        >
-                            <FlatListStyled
-                                data={[
-                                    {
-                                        date: "17/02/2023",
-                                        especiality: "4.500.000/mm³"
-                                    }
-                                ]}
-                                keyExtractor={(item) => {
-                                    const examItem = item as Exams;
-                                    return examItem.date;
-                                }}
-                                renderItem={({ item }) => {
-                                    const examItem = item as Exams;
-                                    const row: Row = {
-                                        date: examItem.date,
-                                        especiality: examItem.especiality
-                                    };
-                                    return <Table rows={[row]} />;
-                                }}
-                            />
-                        </Collapsible>
-
-                        <ListInteractableItem
-                            text="Colesterol Total"
-                            downChevron
-                            changeIcon={collapsedItems.cholesterol}
-                            isButton
-                            modalFunction={() =>
-                                toggleCollapsible("cholesterol")
-                            }
-                        />
-
-                        <Collapsible
-                            collapsed={collapsedItems.cholesterol}
-                            align="center"
-                        >
-                            <FlatListStyled
-                                data={[
-                                    {
-                                        date: "17/02/2023",
-                                        especiality: "25 U/L"
-                                    }
-                                ]}
-                                keyExtractor={(item) => {
-                                    const examItem = item as Exams;
-                                    return examItem.date;
-                                }}
-                                renderItem={({ item }) => {
-                                    const examItem = item as Exams;
-                                    const row: Row = {
-                                        date: examItem.date,
-                                        especiality: examItem.especiality
-                                    };
-                                    return <Table rows={[row]} />;
-                                }}
-                            />
-                        </Collapsible>
-
-                        <ListInteractableItem
-                            text="Glicose"
-                            downChevron
-                            changeIcon={collapsedItems.glucose}
-                            isButton
-                            modalFunction={() => toggleCollapsible("glucose")}
-                        />
-
-                        <Collapsible
-                            collapsed={collapsedItems.glucose}
-                            align="center"
-                        >
-                            <FlatListStyled
-                                data={[
-                                    {
-                                        date: "17/02/2023",
-                                        especiality: "79 mg/dl"
-                                    }
-                                ]}
-                                keyExtractor={(item) => {
-                                    const examItem = item as Exams;
-                                    return examItem.date;
-                                }}
-                                renderItem={({ item }) => {
-                                    const examItem = item as Exams;
-                                    const row: Row = {
-                                        date: examItem.date,
-                                        especiality: examItem.especiality
-                                    };
-                                    return <Table rows={[row]} />;
-                                }}
-                            />
-                        </Collapsible>
-
-                        <ListInteractableItem
-                            text="TGO"
-                            downChevron
-                            changeIcon={collapsedItems.ast}
-                            isButton
-                            modalFunction={() => toggleCollapsible("ast")}
-                        />
-
-                        <Collapsible
-                            collapsed={collapsedItems.ast}
-                            align="center"
-                        >
-                            <FlatListStyled
-                                data={[
-                                    {
-                                        date: "17/02/2023",
-                                        especiality: "25 U/L"
-                                    }
-                                ]}
-                                keyExtractor={(item) => {
-                                    const examItem = item as Exams;
-                                    return examItem.date;
-                                }}
-                                renderItem={({ item }) => {
-                                    const examItem = item as Exams;
-                                    const row: Row = {
-                                        date: examItem.date,
-                                        especiality: examItem.especiality
-                                    };
-                                    return <Table rows={[row]} />;
-                                }}
-                            />
-                        </Collapsible>
-
-                        <ListInteractableItem
-                            text="TGP"
-                            downChevron
-                            changeIcon={collapsedItems.alt}
-                            isButton
-                            modalFunction={() => toggleCollapsible("alt")}
-                        />
-
-                        <Collapsible
-                            collapsed={collapsedItems.alt}
-                            align="center"
-                        >
-                            <FlatListStyled
-                                data={[
-                                    {
-                                        date: "17/02/2023",
-                                        especiality: "28 U/L"
-                                    }
-                                ]}
-                                keyExtractor={(item) => {
-                                    const examItem = item as Exams;
-                                    return examItem.date;
-                                }}
-                                renderItem={({ item }) => {
-                                    const examItem = item as Exams;
-                                    const row: Row = {
-                                        date: examItem.date,
-                                        especiality: examItem.especiality
-                                    };
-                                    return <Table rows={[row]} />;
-                                }}
-                            />
-                        </Collapsible>
-                    </View>
-                    <CancelAndSaveButton
-                        onPress={() => {
-                            setShowExamsDetected(false);
-                            setCollapsedItems({
-                                glucose: true,
-                                hemogram: true,
-                                cholesterol: true,
-                                ast: true,
-                                alt: true
-                            });
-                        }}
-                    />
-                </ModalContainer>
-            ) : (
-                <></>
-            )} */}
         </>
     );
 }
